requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
decorator==4.3.0
fabulous==0.3.0
flake8==3.6.0
flake8-polyfill==1.0.2
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
mccabe==0.6.1
parso==0.3.1
pep8-naming==0.7.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
